align area- and line-chart
