+++ src/index.js
@@ -2,7 +2,6 @@ import AreaChart from './area-chart'
 import StackedAreaChart from './stacked-area-chart'
 import BarChart from './bar-chart'
 import StackedBarChart from './stacked-bar-chart'
-import HorizontalLabeledBarChart from './horizontal-labeled-bar-chart'
 import LineChart from './line-chart'
 import PieChart from './pie-chart'
 import ProgressCircle from './progress-circle'
@@ -24,7 +23,6 @@ export {
     YAxis,
     ProgressCircle,
     WaterfallChart,
-    HorizontalLabeledBarChart,
     Decorators,
     Grid,
     Path,
